# Updates to quest's signed distance example