--- a/src/axom/quest/examples/quest_signed_distance_interface.cpp
+++ b/src/axom/quest/examples/quest_signed_distance_interface.cpp
@@ -68,6 +68,7 @@ struct Arguments
   bool dump_vtk{true};
   bool use_shared{false};
   bool use_batched_query{false};
+  bool ignore_signs{false};
 
   void parse( int argc, char** argv, CLI::App& app)
   {
@@ -105,6 +106,10 @@ struct Arguments
 
     app.add_flag("--use-batched-query", this->use_batched_query, 
       "uses a vectorized query instead of many serial queries")
+      ->capture_default_str();
+
+    app.add_flag("--ignore-signs", this->ignore_signs, 
+      "distance query should ignore signs")
       ->capture_default_str();
 
     app.get_formatter()->column_width(40);
@@ -169,6 +174,7 @@ int main ( int argc, char** argv )
   quest::signed_distance_set_closed_surface( args.is_water_tight );
   quest::signed_distance_set_max_levels( args.maxLevels );
   quest::signed_distance_set_max_occupancy( args.maxOccupancy );
+  quest::signed_distance_set_compute_distance( !args.ignore_signs );
   // _quest_distance_interface_init_start
   int rc = quest::signed_distance_init( args.fileName, global_comm );
   // _quest_distance_interface_init_end