--- a/src/axom/quest/interface/CMakeLists.txt
+++ b/src/axom/quest/interface/CMakeLists.txt
@@ -13,6 +13,7 @@ set(YAML_INPUT_FILE quest_shroud.yaml)
 
 add_shroud(
     YAML_INPUT_FILE      ${YAML_INPUT_FILE}
+    YAML_OUTPUT_DIR      yaml
     C_FORTRAN_OUTPUT_DIR c_fortran
     PYTHON_OUTPUT_DIR    python
 )