--- a/src/axom/quest/interface/python/pyQUESTutil.cpp
+++ b/src/axom/quest/interface/python/pyQUESTutil.cpp
@@ -1,0 +1,8 @@
+// pyQUESTutil.cpp
+// This is generated code, do not edit
+//
+// Copyright (c) 2017-2020, Lawrence Livermore National Security, LLC and
+// other Axom Project Developers. See the top-level COPYRIGHT file for details.
+//
+// SPDX-License-Identifier: (BSD-3-Clause)
+#include "pyQUESTmodule.hpp"