# Summary

- This PR is a feature
- It adds the ability to run the signed distance query example in batched mode
- It also adds an option to `SignedDistance` and its C API to skip the computation of the sign, returning just the magnitude of the distance. This can be useful in applications that do not care about the sign
- It also converts the signed distance example's argument parsing to `CLI11`