@kennyweiss 2020-08-14
@rhornung67 @gzagaris  and others -- thanks for the thorough review.
I think I've addressed all the comments. Please review again at your convenience.