@kennyweiss 2020-08-14
The `batched` option is significantly faster, as expected.
Here's what I got on a full `toss3` node (36 cores):
Without ``--batched``:
```
>./examples/quest_signed_distance_interface_ex -f ../data/quest/sphere.stl --box-dims 100 100 100 
[0][INFO]: initializing signed distance function...
[0][INFO]: input file: ../data/quest/sphere.stl
[0][INFO]: max_levels=15
[0][INFO]: max_occupancy=5
[0][INFO]: time to initialize: 0.00859259s
[0][INFO]: computing mesh bounds...
[0][INFO]: bounding box { min:(-5,-5,-5); max:(5,5,5); range:<10,10,10> }
[0][INFO]: constructing Uniform Mesh of resolution (100,100,100)
[0][INFO]: evaluating signed distance field on specified box mesh...
[0][INFO]: time to evaluate: 24.817s                                   # <--- HERE
[0][INFO]: writing vtk output
```

With ``--batched``:
```
>./examples/quest_signed_distance_interface_ex -f ../data/quest/sphere.stl --box-dims 100 100 100 --batched
[0][INFO]: initializing signed distance function...
[0][INFO]: input file: ../data/quest/sphere.stl
[0][INFO]: max_levels=15
[0][INFO]: max_occupancy=5
[0][INFO]: time to initialize: 0.00853686s
[0][INFO]: computing mesh bounds...
[0][INFO]: bounding box { min:(-5,-5,-5); max:(5,5,5); range:<10,10,10> }
[0][INFO]: constructing Uniform Mesh of resolution (100,100,100)
[0][INFO]: evaluating signed distance field on specified box mesh...
[0][INFO]: time to evaluate batched query: 0.983249s                   # <-- HERE
[0][INFO]: writing vtk output
```